+++ CHANGELOG.rst
@@ -6,7 +6,11 @@ History
 3.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- TextPropagator.extract raises SpanContextCorruptedException with
+  more meaningful error messages.
+
+- Tracer.start_span validates type of ``references=`` preventing
+  problems for users migrating code from opentracing==1.3.0.
 
 
 3.1.0 (2019-05-12)
@@ -100,4 +104,3 @@ History
 ----------------
 
 - Initial public API
-